requirements.txt
# 虚拟环境依赖，保持与 Docker 环境一致
python-telegram-bot==13.15
python-dotenv==0.19.2
notion-client==1.0.0
requests==2.28.2
google-generativeai>=0.3.0
schedule>=1.1.0
python-dateutil>=2.8.2
pytz>=2023.3  # 确保 pytz 已安装
httpx[socks]
PySocks>=1.7.1  # 支持 SOCKS 代理
urllib3==1.26.15  # 与 requests 2.28 兼容的版本
requests[socks]  # 确保 requests 包含 socks 支持
beautifulsoup4==4.10.0
pypdf>=3.0.0
openai>=0.27.0
pyzotero>=1.5.5
markdownify>=0.9.2
firecrawl-py>=1.0.0